Split up utils file
