+++ src/utils/types.ts
@@ -1,0 +1,6 @@
+export interface Hymn {
+    title: string;
+    number: number;
+    content?: string;
+    markdown?: string;
+}
